+++ forms/src/index.ts
@@ -20,7 +20,6 @@ export * from './misc/validators';
 export * from './misc/validationErrorsContainerView';
 export * from './modules/numberInput.module';
 export * from './modules/hasError.module';
-export * from './modules/requiredClass.module';
 export * from './modules/formPipes';
 export * from './modules/errorMessage';
 export * from './misc/formModel';