+++ forms/src/modules/hasError.module.ts
@@ -14,10 +14,10 @@ import {DefaultValidationErrorsComponent, DefaultValidationErrorsContainerCompon
     imports:
     [
         CommonModule,
+        HasErrorDirective,
     ],
     declarations: 
     [
-        HasErrorDirective,
         GroupHasErrorDirective,
         GroupHasErrorContainerDirective,
         ValidationErrorsContainerDirective,